refactor: 우승자 찾는 기능을 Cars로 합침
